pyproject.toml
[project]
name = "ingredionpdfparse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.75.0",
    "docling[vml]>=2.55.1",
    "google-genai==1.53.0",
    "ipython>=8.37.0",
    "langchain>=1.1.0",
    "langchain-core>=1.1.0",
    "langchain-google-genai>=3.2.0",
    "langchain-text-splitters>=1.0.0",
    "pip>=25.3",
    "pymupdf4llm>=0.2.6",
]
